refactor: 구분을 위한 상위 패키지 추가
