+++ src/main/java/com/advanced/app/log/v1/OrderServiceV1.java
@@ -1,16 +1,16 @@
-package com.advanced.app.v3;
+package com.advanced.app.log.v1;
 
 import com.advanced.trace.TraceStatus;
-import com.advanced.trace.logtrace.LogTrace;
+import com.advanced.trace.hellotrace.HelloTraceV1;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class OrderServiceV3 {
+public class OrderServiceV1 {
 
-    private final OrderRepositoryV3 orderRepository;
-    private final LogTrace trace;
+    private final OrderRepositoryV1 orderRepository;
+    private final HelloTraceV1 trace;
 
     public void orderItem(String itemId) {
