+++ src/main/java/com/advanced/app/log/v2/OrderControllerV2.java
@@ -1,30 +1,30 @@
-package com.advanced.app.v3;
+package com.advanced.app.log.v2;
 
 import com.advanced.trace.TraceStatus;
-import com.advanced.trace.logtrace.LogTrace;
+import com.advanced.trace.hellotrace.HelloTraceV2;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-public class OrderControllerV3 {
+public class OrderControllerV2 {
 
-    private final OrderServiceV3 orderService;
-    private final LogTrace trace;
+    private final OrderServiceV2 orderService;
+    private final HelloTraceV2 trace;
 
-    @GetMapping("/v3/request")
+    @GetMapping("/v2/request")
     public String request(String itemId) {
 
         TraceStatus status = null;
         try {
             status = trace.begin("OrderController.request()");
-            orderService.orderItem(itemId);
+            orderService.orderItem(status.getTraceId(), itemId);
             trace.end(status);
             return "ok";
         } catch (Exception e) {
             trace.exception(status, e);
-            throw e;
+            throw e; // 예외가 터진 경우, 예외 로그 출력 후 애플리케이션 흐름이 종료되지 않게 다시 예외를 던져야 한다
         }
     }
 }
